requirements.txt
httplib2==0.11.3
pyserial==3.4
